+++ NEWS.md
@@ -1,4 +1,6 @@
-# FSA 0.8.10 ongoing
+# FSA 0.8.11 ongoing
+* `vbFuns()`: Modified.  Added `Ogle` to list of parameterizations.
+* `vbStarts()`: Modified.  Added the `methLinf=` argument that allows the user to choose if Linf is estimated from a Walford plot (`methLinf="Walford"`; the default and old functionality), as the mean of fish in a certain number of old ages (`methLinf="oldAge"`), or as the mean of a certain number of the longest fish (`methLinf="longFish"`).  The number of ages or long fish is given in `num4Linf=`.  Added methods for `type="Ogle"`.
 
 # FSA 0.8.10 24-Sep-16
 * `alkIndivAge()`: Modified.  Added `na.rm=TRUE` to the checks on the minimum and maximum length data.